pyproject.toml
[tool.poetry]
name = "hh"
version = "1.0.0"
description = ""
authors = ["shyn_le <highlandetspirou2812@gmail.com>"]
readme = "README.md"
packages = [
    { include = "scheduler" },
    { include = "crawler" },
    { include = "singleton" },
]



[tool.poetry.dependencies]
python = ">=3.10,<3.13"
pytest = "^7.4.2"
httpx = "^0.25.0"
beautifulsoup4 = "^4.12.2"
lxml = "^4.9.3"
celery = "^5.3.6"
pymongo = "^4.6.1"
apscheduler = "^3.10.4"
redis = "^5.0.1"
mysql-connector-python = "^8.3.0"
chainexecutor = {git = "https://github.com/highlander-spirou/chain_execute.git"}
neo4j = "^5.18.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"